Version 1.8.0-beta
- Refactor code
- Swap out govips for bimg, rewrite resize function
