+++ .gitignore
@@ -1,8 +1,10 @@
 .DS_Store
-__pycache__
-data
 .vscode
-Go/logs
-Go/cpu.pprof
-Go/go.sum
-Go/config/bot-config.json
+cpu.pprof
+go.sum
+TODO.txt
+logs
+config/bot-config.json
+resize/test-images
+resize/test-images-large
+mem.prof
