--- Go/resize-bot.go
@@ -1,714 +1,0 @@
-/*
-golang rewrite of the telegram sticker resize bot python program.
-*/
-
-package main
-
-import (
-	"bytes"
-	"fmt"
-	"image/png"
-	"io"
-	"io/ioutil"
-	"log"
-	"os"
-	"os/signal"
-	"path/filepath"
-	"sync"
-	"syscall"
-	"time"
-
-	"tg-resize-sticker-images/config"
-	"tg-resize-sticker-images/spam"
-	"tg-resize-sticker-images/stats"
-	"tg-resize-sticker-images/utils"
-
-	"github.com/davidbyttow/govips/v2/vips"
-	"github.com/go-co-op/gocron"
-
-	pngquant "github.com/yusukebe/go-pngquant"
-	tb "gopkg.in/tucnak/telebot.v3"
-)
-
-type Session struct {
-	/* A struct to simplify passing around structs such as utils.Config, SendQueue, bot etc. */
-	bot      *tb.Bot        // Bot this session runs
-	config   *config.Config // Configuration for session
-	spam     *spam.AntiSpam // Anti-spam struct for session
-	queue    *SendQueue     // Message send queue for session
-	lastUser int64          // Keep track of the last user to convert an image
-	Mutex    sync.Mutex     // Avoid concurrent writes
-	vnum     string         // Version
-}
-
-type Message struct {
-	recipient *tb.User       // Recipient of the message
-	bytes     *[]byte        // Photo, as a byte array
-	caption   string         // Caption for the photo
-	sopts     tb.SendOptions // Send options
-}
-
-type SendQueue struct {
-	/* Enforces a rate-limiter to stay within Telegram's send-rate boundaries */
-	messagesPerSecond float32    // Messages-per-second limit
-	messageQueue      []Message  // Queue of messages to send
-	Mutex             sync.Mutex // Mutex to avoid concurrent writes
-}
-
-func addToQueue(queue *SendQueue, message *Message) {
-	queue.Mutex.Lock()
-	queue.messageQueue = append(queue.messageQueue, *message)
-	queue.Mutex.Unlock()
-}
-
-func updateLastUserId(session *Session, id int64) {
-	session.Mutex.Lock()
-	session.lastUser = id
-	session.Mutex.Unlock()
-}
-
-func messageSender(session *Session) {
-	/* Function clears the SendQueue and stays within API limits while doing so */
-	for {
-		// If queue is not empty, clear it
-		if len(session.queue.messageQueue) != 0 {
-			// Lock sendQueue for parsing
-			session.queue.Mutex.Lock()
-
-			// Iterate over queue
-			for i, msg := range session.queue.messageQueue {
-				// If nil bytes, we are only sending text
-				if msg.bytes == nil {
-					_, err := session.bot.Send(msg.recipient, msg.caption, &msg.sopts)
-					if err != nil {
-						log.Println("Error sending non-bytes message in messageSender:", err)
-					}
-				} else {
-					// If non-nil bytes, we are sending a photo
-					sendDocument(session, msg)
-				}
-
-				// Sleep long enough to stay within API limits: convert messagesPerSecond to ms
-				if i < len(session.queue.messageQueue)-1 {
-					time.Sleep(time.Millisecond * time.Duration(1.0/session.queue.messagesPerSecond*1000.0))
-				}
-			}
-
-			// Clear queue
-			session.queue.messageQueue = nil
-
-			// Batch send done, unlock sendQueue
-			session.queue.Mutex.Unlock()
-		}
-
-		// Sleep while waiting for updates
-		time.Sleep(time.Millisecond * 500)
-	}
-}
-
-func resizeImage(imgBytes []byte) (Message, error) {
-	/* Resizes an image in a byte buffer
-
-	Inputs:
-		imgBytes: the image to resize
-
-	Outputs:
-		Message: a message object containing the image and caption
-		error: errors encountered during resize
-	*/
-
-	// Build image from buffer
-	img, err := vips.NewImageFromBuffer(imgBytes)
-	if err != nil {
-		log.Println("⚠️ Error decoding image! Err: ", err)
-
-		errorMsg := fmt.Sprintf("⚠️ Error decoding image (%s).", err.Error())
-		if err.Error() == "unsupported image format" {
-			errorMsg += " Please send JPG/PNG/WebP images."
-		}
-
-		return Message{recipient: nil, bytes: nil, caption: errorMsg}, err
-	}
-
-	// defer closing for later
-	defer img.Close()
-
-	// Dimensions for resize (int)
-	w, h := img.Width(), img.Height()
-
-	// Determine the factor by how much to scale the image with (vips wants f64)
-	var resScale float64
-
-	if w >= h {
-		resScale = 512.0 / float64(w)
-	} else {
-		resScale = 512.0 / float64(h)
-	}
-
-	// Resize, upscale status
-	err = img.Resize(resScale, vips.KernelAuto)
-	imgUpscaled := resScale > 1.0
-
-	if err != nil {
-		log.Println("⚠️ Error resizing image:", err)
-		errorMsg := fmt.Sprintf("⚠️ Error resizing image (%s)", err.Error())
-		return Message{recipient: nil, bytes: nil, caption: errorMsg}, err
-	}
-
-	// Increment compression ratio if size is too large
-	pngParams := vips.PngExportParams{
-		StripMetadata: true,
-		Compression:   6,
-		Interlace:     false,
-	}
-
-	// Encode as png into a new buffer
-	pngBuff, _, err := img.ExportPng(&pngParams)
-	if err != nil {
-		log.Println("⚠️ Error encoding image as png: ", err)
-
-		var errorMsg string
-		if err.Error() == "unsupported image format" {
-			errorMsg = "⚠️ Unsupported image format!"
-		} else {
-			errorMsg = fmt.Sprintf("⚠️ Error encoding image (%s)", err.Error())
-		}
-
-		return Message{recipient: nil, bytes: nil, caption: errorMsg}, err
-	}
-
-	// Did we reach the target file size?
-	compressionFailed := len(pngBuff)/1024 >= 512
-
-	// If compression fails, run the image through pngquant
-	if compressionFailed {
-		expParams := vips.ExportParams{
-			Format:        vips.ImageTypePNG,
-			StripMetadata: true,
-			Compression:   6,
-		}
-
-		imgImg, err := img.ToImage(&expParams)
-		if err != nil {
-			log.Println("⚠️ Error exporting image as image.Image:", err)
-		}
-
-		cImg, err := pngquant.Compress(imgImg, "6")
-		if err != nil {
-			log.Println("⚠️ Error compressing image with pngquant:", err)
-		}
-
-		// Write to buffer
-		cBuff := new(bytes.Buffer)
-		err = png.Encode(cBuff, cImg)
-		if err != nil {
-			log.Println("⚠️ Error encoding cImg as png:", err)
-		}
-
-		pngBuff = cBuff.Bytes()
-		compressionFailed = len(pngBuff)/1024 >= 512
-
-		if compressionFailed {
-			log.Println("\t⚠️ Image compression failed! Buffer length (KB):", len(cBuff.Bytes())/1024)
-		}
-	}
-
-	// Construct the caption
-	imgCaption := fmt.Sprintf(
-		"🖼 Here's your sticker-ready image (%dx%d)! Forward this to @Stickers.",
-		img.Width(), img.Height(),
-	)
-
-	// Add notice to user if image was upscaled or compressed
-	if imgUpscaled {
-		imgCaption += "\n\n⚠️ Image upscaled! Quality may have been lost: consider using a larger image."
-	} else if compressionFailed {
-		imgCaption += "\n\n⚠️ Image compression failed (≥512 KB): you must manually compress the image!"
-	}
-
-	message := Message{recipient: nil, bytes: &pngBuff, caption: imgCaption}
-	return message, nil
-}
-
-func getBytes(session *Session, message *tb.Message, mediaType string) ([]byte, error) {
-	// If using local API, no need to get file: open from disk and return bytes
-	if session.config.API.LocalAPIEnabled {
-		var err error
-		var file tb.File
-
-		// Get file, store
-		switch mediaType {
-		case "photo":
-			file, err = session.bot.FileByID(message.Photo.File.FileID)
-		case "document":
-			file, err = session.bot.FileByID(message.Document.File.FileID)
-		case "sticker":
-			file, err = session.bot.FileByID(message.Sticker.File.FileID)
-		}
-
-		if err != nil {
-			log.Println("⚠️ Error running GetFile (local): ", err)
-			log.Printf("File: %+v\n", file)
-			return []byte{}, err
-		}
-
-		// Construct path from config's working directory
-		fPath := filepath.Join(session.config.API.LocalWorkingDir, session.config.Token, file.FilePath)
-		if err != nil {
-			log.Println("Error creating absolute path:", err)
-			return []byte{}, err
-		}
-
-		// Attempt reading file contents
-		imgBuf, err := ioutil.ReadFile(fPath)
-
-		if err != nil {
-			log.Println("⚠️ Error opening local file from FilePath!", message.Photo.FilePath)
-			log.Println("Constructed fPath:", fPath)
-			log.Println("Err:", err)
-
-			// Error: remove file, return
-			os.Remove(fPath)
-			return []byte{}, err
-		}
-
-		// Success: remove file, return
-		os.Remove(fPath)
-		return imgBuf, nil
-
-	} else {
-		// Else, we're using the regular Telegram bot API: get file from servers
-		var err error
-		var tbFile *tb.File
-		var file io.Reader
-
-		switch mediaType {
-		case "photo":
-			tbFile = message.Photo.MediaFile()
-		case "document":
-			tbFile = message.Document.MediaFile()
-		case "sticker":
-			tbFile = message.Sticker.MediaFile()
-		}
-
-		// Get file
-		file, err = session.bot.File(tbFile)
-
-		if err != nil {
-			log.Println("⚠️ Error running GetFile: ", err)
-			return []byte{}, err
-		}
-
-		// Download or copy to buffer, depending on API used
-		// copy file contents to imgBuf
-		var imgBuf bytes.Buffer
-		_, err = io.Copy(&imgBuf, file)
-
-		if err != nil {
-			log.Println("⚠️ Error copying image to buffer:", err)
-			return []byte{}, err
-		}
-
-		return imgBuf.Bytes(), nil
-	}
-}
-
-func handleIncomingMedia(session *Session, message *tb.Message, mediaType string) {
-	/* Handles incoming media, i.e. those caught by tb.OnPhoto, tb.OnDocument etc. */
-	// Anti-spam: return if user is not allowed to convert
-	if !spam.ConversionPreHandler(session.spam, message.Sender.ID) {
-		log.Println("🚦 Chat", message.Sender.ID, "is ratelimited")
-
-		// Construct message
-		msg := Message{
-			recipient: message.Sender,
-			bytes:     nil,
-			caption:   utils.RatelimitedMessage(session.spam, message.Sender.ID),
-			sopts:     tb.SendOptions{ParseMode: "Markdown"},
-		}
-
-		// Add to send queue
-		go addToQueue(session.queue, &msg)
-		return
-	}
-
-	// Download
-	imgBytes, err := getBytes(session, message, mediaType)
-	if err != nil {
-		// Construct message
-		msg := Message{
-			recipient: message.Sender,
-			bytes:     nil,
-			caption:   "⚠️ Error downloading image! Please try again.",
-		}
-
-		// Add to send queue
-		go addToQueue(session.queue, &msg)
-
-		// Log error, including media type
-		log.Printf("Error downloading image: %s\n", err.Error())
-		return
-	}
-
-	// Resize
-	msg, err := resizeImage(imgBytes)
-	msg.recipient = message.Sender
-
-	// Log errors encountered during resize
-	if err != nil {
-		log.Printf("Error resizing image: %s\n", err.Error())
-	}
-
-	// Add to send queue: regardless of resize outcome, the message is sent
-	go addToQueue(session.queue, &msg)
-
-	// Update stat for count of unique chats in a goroutine
-	if message.Sender.ID != session.lastUser {
-		go stats.UpdateUniqueStat(&message.Sender.ID, session.config)
-		updateLastUserId(session, message.Sender.ID)
-	}
-}
-
-func sendDocument(session *Session, msg Message) {
-	// Send as a document: create object
-	doc := tb.Document{
-		File:     tb.FromReader(bytes.NewReader(*msg.bytes)),
-		Caption:  msg.caption,
-		MIME:     "image/png",
-		FileName: fmt.Sprintf("resized-image-%d.png", time.Now().Unix()),
-	}
-
-	// Disable notifications
-	sendOpts := tb.SendOptions{DisableNotification: true}
-
-	// Send
-	_, err := doc.Send(session.bot, msg.recipient, &sendOpts)
-
-	if err != nil {
-		log.Println("⚠️ Error sending message (notifying user):", err)
-		errorMessage := "🚦 Error sending finished image! Please try again."
-
-		_, err := session.bot.Send(msg.recipient, errorMessage)
-		if err != nil {
-			log.Println("\tUnable to notify user:", err)
-		}
-
-		return
-	}
-
-	go stats.StatsPlusOneConversion(session.config)
-}
-
-func setupBot(session *Session, aspam *spam.AntiSpam) {
-	// Command handler for /start
-	bot := session.bot
-
-	bot.Handle("/start", func(c tb.Context) error {
-		// Anti-spam
-		message := *c.Message()
-		if !spam.CommandPreHandler(aspam, message.Sender.ID, message.Unixtime) {
-			return nil
-		}
-
-		// Construct message
-		startMessage := utils.HelpMessage(&message, aspam)
-		msg := Message{
-			recipient: message.Sender,
-			bytes:     nil,
-			caption:   startMessage,
-			sopts:     tb.SendOptions{ParseMode: "Markdown"},
-		}
-
-		// Add to send queue
-		go addToQueue(session.queue, &msg)
-
-		// Check if the chat is actually new, or just calling /start again
-		if !stats.ChatExists(&message.Sender.ID, session.config) {
-			log.Println("🌟", message.Sender.ID, "bot added to new chat!")
-		}
-
-		return nil
-	})
-
-	// Command handler for /help
-	bot.Handle("/help", func(c tb.Context) error {
-		// Pointer to message
-		message := *c.Message()
-
-		// Anti-spam
-		if !spam.CommandPreHandler(aspam, message.Sender.ID, message.Unixtime) {
-			return nil
-		}
-
-		// Refresh ConversionCount for chat
-		spam.RefreshConversions(aspam, message.Sender.ID)
-
-		// Help message
-		helpMessage := utils.HelpMessage(&message, aspam)
-
-		// Construct message
-		msg := Message{
-			recipient: message.Sender,
-			bytes:     nil,
-			caption:   helpMessage,
-			sopts:     tb.SendOptions{ParseMode: "Markdown"},
-		}
-
-		// Add to send queue
-		go addToQueue(session.queue, &msg)
-
-		if message.Sender.ID != session.config.Owner {
-			log.Println("🙋", message.Sender.ID, "requested help!")
-		}
-
-		return nil
-	})
-
-	// Command handler for /stats
-	bot.Handle("/stats", func(c tb.Context) error {
-		// Pointer to message
-		message := *c.Message()
-
-		// Anti-spam
-		if !spam.CommandPreHandler(aspam, message.Sender.ID, message.Unixtime) {
-			return nil
-		}
-
-		// Get stats message
-		caption, sopts := stats.BuildStatsMsg(session.config, aspam, session.vnum)
-
-		// Construct message
-		msg := Message{recipient: message.Sender, bytes: nil, caption: caption, sopts: sopts}
-
-		// Add to send queue
-		go addToQueue(session.queue, &msg)
-
-		if message.Sender.ID != session.config.Owner {
-			log.Println("📊", message.Sender.ID, "requested to view stats!")
-		}
-
-		return nil
-	})
-
-	// Register photo handler
-	bot.Handle(tb.OnPhoto, func(c tb.Context) error {
-		// Pointer to message
-		message := *c.Message()
-
-		handleIncomingMedia(session, &message, "photo")
-		return nil
-	})
-
-	// Register document handler
-	bot.Handle(tb.OnDocument, func(c tb.Context) error {
-		// Pointer to message
-		message := *c.Message()
-
-		handleIncomingMedia(session, &message, "document")
-		return nil
-	})
-
-	// Register sticker handler
-	bot.Handle(tb.OnSticker, func(c tb.Context) error {
-		// Pointer to message
-		message := *c.Message()
-
-		handleIncomingMedia(session, &message, "sticker")
-		return nil
-	})
-
-	// Register handler for incoming callback queries (i.e. stats refresh)
-	bot.Handle(tb.OnCallback, func(c tb.Context) error {
-		cb := *c.Callback()
-
-		// Anti-spam
-		if !spam.CommandPreHandler(aspam, cb.Sender.ID, time.Now().Unix()) {
-			resp := tb.CallbackResponse{
-				CallbackID: cb.ID,
-				Text:       "⚠️ Please do not spam the bot for no reason.",
-				ShowAlert:  true,
-			}
-
-			bot.Respond(&cb, &resp)
-			return nil
-		}
-
-		if cb.Data == "stats/refresh" {
-			msg, sopts := stats.BuildStatsMsg(session.config, aspam, session.vnum)
-			bot.Edit(cb.Message, msg, &sopts)
-
-			resp := tb.CallbackResponse{
-				CallbackID: cb.ID,
-				Text:       "🔄 Statistics refreshed!",
-				ShowAlert:  false,
-			}
-
-			bot.Respond(&cb, &resp)
-		} else {
-			log.Println("⚠️ Invalid callback data received:", cb.Data)
-		}
-
-		return nil
-	})
-}
-
-func setupSignalHandler(conf *config.Config) {
-	// Listens for incoming interrupt signals, dumps config if detected
-	channel := make(chan os.Signal)
-	signal.Notify(channel, os.Interrupt, syscall.SIGTERM, syscall.SIGINT)
-
-	go func() {
-		<-channel
-		log.Println("🚦 Received interrupt signal: dumping config...")
-		config.DumpConfig(conf)
-		os.Exit(0)
-	}()
-}
-
-func main() {
-	/* Version history
-	0.0.0: 2021.03.29: started
-	1.0.0: 2021.05.15: first go implementation
-	1.1.0: 2021.05.16: keeping track of unique chats, binsearch
-	1.2.0: 2021.05.17: callback buttons for /stats
-	1.3.0: 2021.05.17: image compression with pngquant
-	1.3.1: 2021.05.19: bug fixes, error handling
-	1.4.0: 2021.08.22: error handling, local API support, handle interrupts
-	1.4.1: 2021.08.25: logging changes to reduce disk writes
-	1.5.0: 2021.08.30: added anti-spam measures, split the program into modules
-	1.5.1: 2021.09.01: fix concurrent map writes
-	1.5.2: 2021.09.09: improvements to spam management
-	1.5.3: 2021.09.10: address occasional runtime errors
-	1.5.4: 2021.09.13: tweaks to file names
-	1.5.5: 2021.09.15: tweaks to error messages, memory
-	1.5.6: 2021.09.27: logging improvements, add anti-spam insights
-	1.5.7: 2021.09.30: callbacks for /spam, logging
-	1.5.8: 2021.11.11: improvements to /spam command, bump telebot + core
-	1.6.0: 2021.11.13: implement a message send queue, locks for config
-	1.6.1: 2021.11.13: send error messages with queue
-	1.6.2: 2021.11.14: add session struct, simplify media handling, add webp support
-	1.6.3: 2021.11.15: log dl/resize failures, improve /start
-	1.6.4: 2021.11.15: don't store chat ID on /start
-	1.7.0: 2021.12.08: upgrade to telebot v3 and migrate code
-	1.7.1: 2021.12.21: code refactor, bump deps */
-	const vnum string = "1.7.1 (2021.12.21)"
-
-	// Log file
-	wd, _ := os.Getwd()
-	logPath := filepath.Join(wd, "logs")
-	if _, err := os.Stat(logPath); os.IsNotExist(err) {
-		_ = os.Mkdir(logPath, os.ModePerm)
-	}
-
-	// Set-up logging
-	logFilePath := filepath.Join(logPath, "bot-log.log")
-	logf, err := os.OpenFile(logFilePath, os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0644)
-	if err != nil {
-		log.Println(err)
-	}
-
-	// Set output of logs to f
-	defer logf.Close()
-	log.SetOutput(logf)
-	log.Printf("🤖 [%s] Bot started at %s", vnum, time.Now())
-
-	// Load (or create) config
-	conf := config.LoadConfig()
-
-	// Setup anti-spam
-	Spam := spam.AntiSpam{}
-	Spam.ChatBannedUntilTimestamp = make(map[int64]int64)
-	Spam.ChatConversionLog = make(map[int64]spam.ConversionLog)
-	Spam.ChatBanned = make(map[int64]bool)
-	Spam.Rules = make(map[string]int64)
-
-	// Add rules
-	Spam.Rules["ConversionsPerHour"] = conf.ConversionRate
-	Spam.Rules["TimeBetweenCommands"] = 2
-
-	// Setup signal handler
-	setupSignalHandler(conf)
-
-	// Verify we're logged out if we're using the cloud API
-	if conf.API.LocalAPIEnabled && !conf.API.CloudAPILoggedOut {
-		log.Println("🚦 Local bot API enabled: logging out from cloud API servers...")
-
-		// Start bot in regular mode
-		bot, err := tb.NewBot(tb.Settings{
-			URL:    "https://api.telegram.org",
-			Token:  conf.Token,
-			Poller: &tb.LongPoller{Timeout: 10 * time.Second},
-		})
-
-		if err != nil {
-			log.Println("Error starting bot during logout:", err)
-			return
-		}
-
-		// Logout from the cloud API server
-		success, err := bot.Logout()
-
-		if success {
-			log.Println("✅ Successfully logged out from the cloud API server!")
-		} else {
-			log.Println("⚠️ Error logging out from the server:", err)
-			return
-		}
-
-		// Success: update config, dump
-		conf.API.CloudAPILoggedOut = true
-		config.DumpConfig(conf)
-
-		log.Println("✅ Config updated to use local API!")
-
-		// Warn if working directory is unset
-		if conf.API.LocalWorkingDir == "working/dir/on/server" || conf.API.LocalWorkingDir == "" {
-			log.Fatal("⚠️ Local API is enabled, but LocalWorkingDir is unset! Images cannot be downloaded.")
-		}
-
-		fmt.Println("✅ Logged out from cloud API server: please restart the program.")
-		os.Exit(0)
-	}
-
-	// Create bot
-	bot, err := tb.NewBot(tb.Settings{
-		URL:    conf.API.URL,
-		Token:  conf.Token,
-		Poller: &tb.LongPoller{Timeout: 10 * time.Second},
-	})
-
-	if err != nil {
-		log.Fatal("Error starting bot:", err)
-		return
-	}
-
-	// https://pkg.go.dev/github.com/davidbyttow/govips/v2@v2.6.0/vips#LoggingSettings
-	vips.LoggingSettings(nil, vips.LogLevel(3))
-	vips.Startup(nil)
-
-	// Setup messageSender
-	sendQueue := SendQueue{messagesPerSecond: 30.0}
-
-	// Define session: used to throw around structs that are needed frequently
-	session := Session{
-		bot:    bot,
-		config: conf,
-		spam:   &Spam,
-		queue:  &sendQueue,
-		vnum:   vnum,
-	}
-
-	// Run messageSender in a goroutine
-	go messageSender(&session)
-
-	// Setup bot
-	setupBot(&session, &Spam)
-
-	// Dump statistics to disk once every 30 minutes, clean spam struct every 60 minutes
-	scheduler := gocron.NewScheduler(time.UTC)
-	scheduler.Every(30).Minutes().Do(config.DumpConfig, conf)
-	scheduler.Every(60).Minutes().Do(spam.CleanConversionLogs, &Spam)
-	scheduler.StartAsync()
-
-	session.bot.Start()
-}
