+++ README.md
@@ -1,55 +1,35 @@
 # tg-resize-sticker-images
-A tiny, privacy-preserving bot that converts and compresses images so that they can be turned into Telegram stickers. Reachable on Telegram as [@resizeimgforstickerbot](https://t.me/resizeimgforstickerbot).
+A fast and tiny, privacy-preserving bot that converts and compresses images so that they can be turned into Telegram stickers. Reachable on Telegram as [@resizeimgforstickerbot](https://t.me/resizeimgforstickerbot).
 
-This repository contains both the initial Python implementation and a newer Golang implementation. The bot currently runs the version written in Go; the version can be checked with the `/stats` command.
-
-There are some slight differences between the two implementations:
-
-**Python**
-- image conversion and compression through pillow
-- Redis for statistics
-
-**Go**
-- image conversion through vips
+**Bot features**
+- fast image conversion, compression, and sending due to in-memory processing
+- image conversion through libvips
 - image compression through pngquant
 - statistics periodically dumped from memory to a json-file
 
-Both handle images exclusively in memory, and _do not_ store or cache received files. In order to collect statistics on how many people use the bot, the user ID of every user is stored in a json-file. This is the only information collected.
+The bot handles images exclusively in memory, and _does not_ store or cache received files. In order to collect statistics on how many people use the bot, the user ID of every user is stored in a json-file. This is the only information collected.
+
+The current version the bot runs can be seen by running the `/stats` command.
 
 ## Compiling
-Compiling the program from source requires [vips](https://libvips.github.io/libvips/), which can be found from most package managers, including apt and homebrew. With vips installed, run `git clone https://github.com/499602D2/tg-resize-sticker-images`, cd into `tg-resize-sticker-images/Go/src` and run `go build -o resize-bot`. Now you can run the program with `./resize-bot`. The program creates two folders; `config` and `logs`.
+Compiling the program from source requires [vips](https://www.libvips.org/, which can be found in most package managers, including apt and homebrew. With vips installed, run `git clone https://github.com/499602D2/tg-resize-sticker-images`, cd into `/tg-resize-sticker-images` and run `go build -o resize-bot`. Now you can run the program with `./resize-bot`. The program stores log-files under `logs`.
 
-### Possible compilation errors
+### Possible compilation errors (macOS)
     go build github.com/davidbyttow/govips/v2/vips: invalid flag in pkg-config --cflags: -Xpreprocessor
 
 This can be fixed by running `export CGO_CFLAGS_ALLOW=-Xpreprocessor` in your shell.
 
 
-## Configuration of the Go version: Cloud API server (default)
-Configuration is stored in `botConfig.json`, under the `Go/src/config` folder. If running the program without a local bot API server, the setup is trivial: you're asked to enter your bot's API key, and then you're ready to go.
+## Configuring the bot
+Configuration is stored in `botConfig.json`, under the `/config` folder. The setup is trivial: you're asked to enter your bot's API key, and then you're ready to go.
 
-Additionally, you can set the `Owner` property in the configuration file to your own Telegram user's ID, in order to disable logging of requests made from said account. Finding out your user ID should be trivial from the logs: simply convert an image or run a command.
+Additionally, you can set the `Owner` property in the configuration file to your own Telegram user ID, in order to disable logging of requests made from said account. Finding out your user ID should be trivial from the logs: simply convert an image or run a command.
 
-## Configuration of the Go version: local bot API server
-**Warning:** using a local bot API server is currently not recommended, as it does not support memory-only operations like the cloud API does. This means that every image is downloaded to disk, instead of simply being downloaded temporarily to memory.
-
-If you're using a local bot API server, the configuration is a little more nuanced. To convert from the default cloud API servers to a local server, first start the bot normally. Then, open `botConfig.json` and under `API` set `LocalAPIEnabled` to `true` and `URL` to your API server's endpoint (default `http://localhost:8081`).
-
-Now start the program once: verify that you see the text `✅ Logged out from cloud API server: please restart the program.` in the log files. Then, open the configuration file again and verify that `CloudAPILoggedOut` is set to `True`, assuming the log-out was successful.
-
-Now, add in your bot API server's working directory into `LocalWorkingDir`: this should be _relative_ to the path of the program. As an example, assuming you have compiled the program and have the compiled as e.g. `resize-bot` in the folder `~/tg-resize-sticker-images/Go/src` and the API server running under `~/telegram-bot-api`, you would set `LocalWorkingDir` to `../../../telegram-bot-api`.
-
-Sample configuration could look as follows (when the executable resides in `~/tg-resize-sticker-images/Go/src`):
+A sample configuration file looks as follows:
 
 ```
 {
     "Token": "12345:abcdefgh",
-    "API": {
-        "LocalAPIEnabled": true,
-	"CloudAPILoggedOut": true,
-	"LocalWorkingDir": "../../../telegram-bot-api",
-	"URL": "http://localhost:8081"
-    },
     "Owner": 12345,
     "StatConverted": 10,
     "StatUniqueChats": 2,
@@ -60,3 +40,33 @@ Sample configuration could look as follows (when the executable resides in `~/tg
     ]
 }
 ```
+
+## Changelog
+<details>
+  <summary>View historical changelog</summary>
+	0.0.0: 2021.03.29: started
+	1.0.0: 2021.05.15: first go implementation
+	1.1.0: 2021.05.16: keeping track of unique chats, binsearch
+	1.2.0: 2021.05.17: callback buttons for /stats
+	1.3.0: 2021.05.17: image compression with pngquant
+	1.3.1: 2021.05.19: bug fixes, error handling
+	1.4.0: 2021.08.22: error handling, local API support, handle interrupts
+	1.4.1: 2021.08.25: logging changes to reduce disk writes
+	1.5.0: 2021.08.30: added anti-spam measures, split the program into modules
+	1.5.1: 2021.09.01: fix concurrent map writes
+	1.5.2: 2021.09.09: improvements to spam management
+	1.5.3: 2021.09.10: address occasional runtime errors
+	1.5.4: 2021.09.13: tweaks to file names
+	1.5.5: 2021.09.15: tweaks to error messages, memory
+	1.5.6: 2021.09.27: logging improvements, add anti-spam insights
+	1.5.7: 2021.09.30: callbacks for /spam, logging
+	1.5.8: 2021.11.11: improvements to /spam command, bump telebot + core
+	1.6.0: 2021.11.13: implement a message send queue, locks for config
+	1.6.1: 2021.11.13: send error messages with queue
+	1.6.2: 2021.11.14: add session struct, simplify media handling, add webp support
+	1.6.3: 2021.11.15: log dl/resize failures, improve /start
+	1.6.4: 2021.11.15: don't store chat ID on /start
+	1.7.0: 2021.12.08: upgrade to telebot v3 and migrate code
+	1.7.1: 2021.12.21: code refactor, bump deps
+	1.8.0: 2022.02.01: remove local API code, refactor, small fixes, rewrite resize function
+</details>
