+++ go.mod
@@ -3,18 +3,19 @@ module tg-resize-sticker-images
 go 1.17
 
 require (
-	github.com/davidbyttow/govips/v2 v2.8.1
+	github.com/davidbyttow/govips/v2 v2.10.0
 	github.com/dustin/go-humanize v1.0.0
 	github.com/go-co-op/gocron v1.11.0
 	github.com/pkg/errors v0.9.1 // indirect
 	github.com/yusukebe/go-pngquant v0.0.0-20200223090257-49b91f11b627
 	golang.org/x/image v0.0.0-20211028202545-6944b10bf410 // indirect
-	golang.org/x/net v0.0.0-20211216030914-fe4d6282115f // indirect
+	golang.org/x/net v0.0.0-20220127200216-cd36cc0744dd // indirect
 	golang.org/x/text v0.3.7 // indirect
-	gopkg.in/tucnak/telebot.v3 v3.0.0-20211217125808-d053f01c3002
+	gopkg.in/telebot.v3 v3.0.0-20220130115853-f0291132d3c3
 )
 
 require (
+	github.com/h2non/bimg v1.1.6 // indirect
 	github.com/robfig/cron/v3 v3.0.1 // indirect
 	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c // indirect
 )
